refactor(web): move ChatEvent out of lib/ into shared/ (#535 follow-up 4)
apps/web/src/lib/ is reserved for browser-only utilities per the 3-tier
architecture rules, but ChatEvent is the SSE wire schema shared by both
halves. Two server-tier modules (sse-writer, jsonl-message-to-events)
were importing from ../../lib/chat-events, breaking the doc's
"server tier does not depend on lib/" rule.

Moved apps/web/src/lib/chat-events.ts to apps/web/src/shared/chat-events.ts
and updated all 8 importers (api, server services, components, tests).

Refs: https://github.com/band-app/band/issues/535
