+++ apps/web/src/components/chat/use-chat-subscription.ts
@@ -34,7 +34,7 @@
  */
 
 import { useCallback, useEffect, useReducer, useRef, useState } from "react";
-import { CHAT_EVENT_TYPES, type ChatEvent, type ChatEventFile } from "../../lib/chat-events";
+import { CHAT_EVENT_TYPES, type ChatEvent, type ChatEventFile } from "../../shared/chat-events";
 import { trpc } from "../../lib/trpc-client";
 import { type ChatSubscriptionState, chatEventReducer, INITIAL_STATE } from "./chat-event-reducer";
