+++ apps/web/src/server/services/jsonl-message-to-events.ts
@@ -16,7 +16,7 @@
  * the same pattern (`chat-event-reducer.ts`).
  */
 
-import type { ChatEvent } from "../../lib/chat-events";
+import type { ChatEvent } from "../../shared/chat-events";
 
 /** Minimal shape of a persisted message we accept — matches
  *  `SessionMessageItem` from `@band-app/coding-agent` but kept local
